Added samples.
Also fixed problem related to Gizmos drawing lines in the base color when a specific color is passed.
Added a range scale for a FloatModifier, making the increment per drag relative to the size of the bounds (it's more usable when using large ranges and less obnoxious when using small ranges).